refactor: created a new package for User in domain layer
