+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/domain/model/user/User.kt
@@ -1,4 +1,4 @@
-package br.com.igorrodrigues.farmcontrol.domain.model
+package br.com.igorrodrigues.farmcontrol.domain.model.user
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import javax.persistence.*
