+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/domain/usecase/user/CreateUserUseCase.kt
@@ -1,7 +1,7 @@
 package br.com.igorrodrigues.farmcontrol.domain.usecase.user
 
-import br.com.igorrodrigues.farmcontrol.domain.model.AllUser
-import br.com.igorrodrigues.farmcontrol.domain.model.User
+import br.com.igorrodrigues.farmcontrol.domain.model.user.AllUser
+import br.com.igorrodrigues.farmcontrol.domain.model.user.User
 import org.springframework.stereotype.Service
 
 @Service
@@ -12,6 +12,7 @@ class CreateUserUseCase(private val allUser: AllUser,
     fun create(userDto: UserDto): User {
         if (allUser.userAlreadyExist(userDto.email)) throw UserAlreadyExistentException()
         val saved = allUser.save(User(email = userDto.email, password = userDto.password))
+        //TODO: Maybe tenantService could be handled in persistence layer instead here.
         tenantService.initDatabase(saved.email)
         return saved
     }
