+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/infrastructure/persistence/AllUserRepository.kt
@@ -1,6 +1,6 @@
 package br.com.igorrodrigues.farmcontrol.infrastructure.persistence
 
-import br.com.igorrodrigues.farmcontrol.domain.model.User
+import br.com.igorrodrigues.farmcontrol.domain.model.user.User
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface AllUserRepository: JpaRepository<User, Long> {
