+++ src/test/kotlin/br/com/igorrodrigues/farmcontrol/controller/auth/AuthControllerTest.kt
@@ -1,7 +1,7 @@
 package br.com.igorrodrigues.farmcontrol.controller.auth
 
 import br.com.igorrodrigues.farmcontrol.controller.ErrorDetail
-import br.com.igorrodrigues.farmcontrol.domain.model.User
+import br.com.igorrodrigues.farmcontrol.domain.model.user.User
 import br.com.igorrodrigues.farmcontrol.domain.usecase.user.CreateUserUseCase
 import br.com.igorrodrigues.farmcontrol.domain.usecase.user.UserDto
 import br.com.igorrodrigues.farmcontrol.infrastructure.security.TokenDto
