+++ src/test/kotlin/br/com/igorrodrigues/farmcontrol/infrastructure/persistence/AllUserRepositoryTest.kt
@@ -1,6 +1,6 @@
 package br.com.igorrodrigues.farmcontrol.infrastructure.persistence
 
-import br.com.igorrodrigues.farmcontrol.domain.model.User
+import br.com.igorrodrigues.farmcontrol.domain.model.user.User
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.springframework.beans.factory.annotation.Autowired
